Build: Revert to ScalaTest
It seems that MUnit does not support JS testing with dependencies, or something like that. It has been a while, I forget. Find my issue in their repo, if context is needed

I am not importing ScalaTest Matchers anymore because it made autocomplete impossible in IntelliJ, it was polluted with a billion should* methods.